Move EnrollmentPanel to its own file
